pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "theteam"
dynamic = ["version"]
description = "Local-first LLM agent coordination and development suite"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["llm", "agent", "ai", "workflow", "coordination"]
authors = [
  { name = "Dane Howard", email = "mirrord@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "ollama>=0.1.0",
  "pyyaml>=6.0",
  "networkx>=3.0",
  "flask>=3.0.0",
  "flask-socketio>=5.3.0",
  "flask-cors>=4.0.0",
  "python-socketio>=5.11.0",
  "eventlet>=0.35.0",
  "chromadb>=0.4.0",
]

[project.optional-dependencies]
test = [
  "pytest>=7.0",
  "pytest-benchmark>=4.0",
]
dev = [
  "mypy>=1.0.0",
]
benchmark = [
  "pandas>=2.0.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.12.0",
  "tqdm>=4.65.0",
]

[project.scripts]
pithos = "pithos.__main__:main"
pithos-agent = "pithos.agent:main"
pithos-config = "pithos.config_manager:main"
pithos-conditions = "pithos.conditions:main"
pithos-tools = "pithos.tools:main"
pithos-memory = "pithos.memory_tool:main"
pithos-database = "pithos.database_manager:main"
pithos-flowcharts = "pithos.flowchart_store:main"
pithos-benchmark = "benchmarks.easy_problems.main:main"
theteam = "theteam.server:main"

[project.urls]
Documentation = "https://github.com/mirrord/theteam#readme"
Issues = "https://github.com/mirrord/theteam/issues"
Source = "https://github.com/mirrord/theteam"

[tool.hatch.version]
path = "src/theteam/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/theteam tests}"

[tool.coverage.run]
source_pkgs = ["theteam", "tests"]
branch = true
parallel = true
omit = [
  "src/theteam/__about__.py",
]

[tool.coverage.paths]
theteam = ["src/theteam", "*/theteam/src/theteam"]
tests = ["tests", "*/theteam/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
